gacha system fix

+++ UnlimitedTowerContract/Contract_eosiocpp/Rule/rule_item_id_table.hpp
@@ -11,12 +11,12 @@ struct item_stat
 class critemid
 {
   public:
-    uint64_t i_id;
+    uint8_t i_id;
     uint8_t i_type;
     item_stat i_min_range;
     item_stat i_max_range;
 public:
-    uint64_t primary_key() const {return i_id;}
+    uint8_t primary_key() const {return i_id;}
 };
 
 typedef multi_index<N(critemid),critemid> item_id_data_table;
